+++ opensquirrel/circuit.py
@@ -2,15 +2,10 @@ from __future__ import annotations
 
 from collections.abc import Callable
 from typing import TYPE_CHECKING, Any
-
-import numpy as np
-
 from opensquirrel.passes.exporter import ExportFormat
 
 if TYPE_CHECKING:
-    from numpy.typing import NDArray
-
-    from opensquirrel.ir import IR, Gate, QubitLike
+    from opensquirrel.ir import IR, Gate
     from opensquirrel.passes.decomposer import Decomposer
     from opensquirrel.passes.mapper import Mapper
     from opensquirrel.passes.merger import Merger
@@ -47,7 +42,6 @@ class Circuit:
         """Create a circuit object from a register manager and an IR."""
         self.register_manager = register_manager
         self.ir = ir
-        self.phase_map = self.PhaseMap(np.zeros(self.qubit_register_size, dtype=np.complex128))
 
     def __repr__(self) -> str:
         """Write the circuit to a cQASM 3 string."""
@@ -111,7 +105,7 @@ class Circuit:
         """
         from opensquirrel.passes.decomposer import general_decomposer
 
-        general_decomposer.decompose(self, decomposer)
+        general_decomposer.decompose(self.ir, decomposer)
 
     def map(self, mapper: Mapper) -> None:
         """Generic qubit mapper pass.
@@ -128,7 +122,7 @@ class Circuit:
         """
         from opensquirrel.passes.decomposer import general_decomposer
 
-        general_decomposer.replace(self, gate, replacement_gates_function)
+        general_decomposer.replace(self.ir, gate, replacement_gates_function)
 
     def export(self, fmt: ExportFormat | None = None) -> Any:
         if fmt == ExportFormat.QUANTIFY_SCHEDULER:
@@ -141,22 +135,3 @@ class Circuit:
             return cqasmv1_exporter.export(self)
         msg = "unknown exporter format"
         raise ValueError(msg)
-
-    class PhaseMap:
-        def __init__(self, phase_map: NDArray[np.complex128]) -> None:
-            """Initialize a PhaseMap object."""
-            self.qubit_phase_map = phase_map
-
-        def __contains__(self, qubit: QubitLike) -> bool:
-            """Checks if qubit is in the phase map."""
-            return qubit in self.qubit_phase_map
-
-        def add_qubit_phase(self, qubit: QubitLike, phase: np.complex128) -> None:
-            from opensquirrel.ir import Qubit
-
-            self.qubit_phase_map[Qubit(qubit).index] += phase
-
-        def get_qubit_phase(self, qubit: QubitLike) -> np.complex128:
-            from opensquirrel.ir import Qubit
-
-            return np.complex128(self.qubit_phase_map[Qubit(qubit).index])