+++ opensquirrel/passes/merger/general_merger.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from math import cos, floor, log10, sin
-from typing import cast
+from typing import Any, cast
 
 import numpy as np
 
@@ -20,13 +20,14 @@ def compose_bloch_sphere_rotations(bsr_a: BlochSphereRotation, bsr_b: BlochSpher
         A q
         B q
 
-    A linear operations:
+    As linear operations:
         (B * A) q
 
     If the final Bloch sphere rotation is anonymous, we try to match it to a default gate.
 
     Uses Rodrigues' rotation formula (see https://en.wikipedia.org/wiki/Rodrigues%27_rotation_formula).
     """
+
     if bsr_a.qubit != bsr_b.qubit:
         msg = "cannot merge two Bloch sphere rotations on different qubits"
         raise ValueError(msg)
@@ -129,6 +130,8 @@ def rearrange_barriers(ir: IR) -> None:
 
 
 class Merger(ABC):
+    def __init__(self, **kwargs: Any) -> None: ...
+
     @abstractmethod
     def merge(self, ir: IR, qubit_register_size: int) -> None:
         raise NotImplementedError