+++ opensquirrel/passes/router/__init__.py
@@ -1,3 +1,5 @@
-from opensquirrel.passes.router.general_router import Router
+from opensquirrel.passes.router.general_router import Router  # noqa: I001
+from opensquirrel.passes.router.astar_router import AStarRouter
+from opensquirrel.passes.router.shortest_path_router import ShortestPathRouter
 
-__all__ = ["Router"]
+__all__ = ["AStarRouter", "Router", "ShortestPathRouter"]