+++ opensquirrel/writer/writer.py
@@ -6,10 +6,12 @@ from opensquirrel.ir import (
     CR,
     CZ,
     SWAP,
+    AsmDeclaration,
     Barrier,
     Bit,
     BlochSphereRotation,
     BsrAngleParam,
+    BsrFullParams,
     BsrNoParams,
     ControlledGate,
     CRk,
@@ -22,6 +24,8 @@ from opensquirrel.ir import (
     Measure,
     Qubit,
     Reset,
+    String,
+    SupportsStr,
     Wait,
 )
 from opensquirrel.register_manager import RegisterManager
@@ -45,6 +49,10 @@ class _WriterImpl(IRVisitor):
             (f"bit[{bit_register_size}] {bit_register_name}" if bit_register_size > 0 else ""),
         )
 
+    def visit_str(self, s: SupportsStr) -> str:
+        s = String(s)
+        return f"{s.value}"
+
     def visit_float(self, f: SupportsFloat) -> str:
         f = Float(f)
         return f"{f.value:.{self.FLOAT_PRECISION}}"
@@ -61,6 +69,11 @@ class _WriterImpl(IRVisitor):
         qubit_register_name = self.register_manager.get_qubit_register_name()
         return f"{qubit_register_name}[{qubit.index}]"
 
+    def visit_asm_declaration(self, asm_declaration: AsmDeclaration) -> None:
+        backend_name = asm_declaration.backend_name.accept(self)
+        backend_code = asm_declaration.backend_code.accept(self)
+        self.output += f"asm({backend_name}) '''{backend_code}'''\n"
+
     def visit_gate(self, gate: Gate) -> None:
         if gate.name == "Gate":
             self.output += f"{gate}\n"
@@ -72,6 +85,15 @@ class _WriterImpl(IRVisitor):
     def visit_bsr_no_params(self, gate: BsrNoParams) -> None:
         qubit_operand = gate.qubit.accept(self)
         self.output += f"{gate.name} {qubit_operand}\n"
+
+    def visit_bsr_full_params(self, gate: BsrFullParams) -> None:
+        nx = gate.nx.accept(self)
+        ny = gate.ny.accept(self)
+        nz = gate.nz.accept(self)
+        theta_argument = gate.theta.accept(self)
+        phi_argument = gate.phi.accept(self)
+        qubit_operand = gate.qubit.accept(self)
+        self.output += f"{gate.name}({nx}, {ny}, {nz}, {theta_argument}, {phi_argument}) {qubit_operand}\n"
 
     def visit_bsr_angle_param(self, gate: BsrAngleParam) -> None:
         theta_argument = gate.theta.accept(self)